Extract common components
